+++ sdk/resourcemanager/compute/armcompute/zz_generated_models.go
@@ -10,10 +10,9 @@ package armcompute
 
 import (
 	"encoding/json"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"reflect"
 	"time"
-
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 )
 
 // APIEntityReference - The API entity reference.
@@ -233,11 +232,6 @@ func (a AvailabilitySetUpdate) MarshalJSON() ([]byte, error) {
 	populate(objectMap, "properties", a.Properties)
 	populate(objectMap, "sku", a.SKU)
 	return json.Marshal(objectMap)
-}
-
-// AvailabilitySetsCreateOrUpdateOptions contains the optional parameters for the AvailabilitySets.CreateOrUpdate method.
-type AvailabilitySetsCreateOrUpdateOptions struct {
-	// placeholder for future optional parameters
 }
 
 // AvailabilitySetsDeleteOptions contains the optional parameters for the AvailabilitySets.Delete method.
@@ -1268,6 +1262,184 @@ type CloudServicesUpdateDomainListUpdateDomainsOptions struct {
 	// placeholder for future optional parameters
 }
 
+// CommunityGalleriesGetOptions contains the optional parameters for the CommunityGalleries.Get method.
+type CommunityGalleriesGetOptions struct {
+	// placeholder for future optional parameters
+}
+
+// CommunityGallery - Specifies information about the Community Gallery that you want to create or update.
+type CommunityGallery struct {
+	PirCommunityGalleryResource
+}
+
+// CommunityGalleryIdentifier - The identifier information of community gallery.
+type CommunityGalleryIdentifier struct {
+	// The unique id of this community gallery.
+	UniqueID *string `json:"uniqueId,omitempty"`
+}
+
+// CommunityGalleryImage - Specifies information about the gallery image definition that you want to create or update.
+type CommunityGalleryImage struct {
+	PirCommunityGalleryResource
+	// Describes the properties of a gallery image definition.
+	Properties *CommunityGalleryImageProperties `json:"properties,omitempty"`
+}
+
+// CommunityGalleryImageProperties - Describes the properties of a gallery image definition.
+type CommunityGalleryImageProperties struct {
+	// REQUIRED; This is the gallery image definition identifier.
+	Identifier *GalleryImageIdentifier `json:"identifier,omitempty"`
+
+	// REQUIRED; This property allows the user to specify whether the virtual machines created under this image are 'Generalized' or 'Specialized'.
+	OSState *OperatingSystemStateTypes `json:"osState,omitempty"`
+
+	// REQUIRED; This property allows you to specify the type of the OS that is included in the disk when creating a VM from a managed image.
+	// Possible values are:
+	// Windows
+	// Linux
+	OSType *OperatingSystemTypes `json:"osType,omitempty"`
+
+	// Describes the disallowed disk types.
+	Disallowed *Disallowed `json:"disallowed,omitempty"`
+
+	// The end of life date of the gallery image definition. This property can be used for decommissioning purposes. This property is updatable.
+	EndOfLifeDate *time.Time `json:"endOfLifeDate,omitempty"`
+
+	// A list of gallery image features.
+	Features []*GalleryImageFeature `json:"features,omitempty"`
+
+	// The hypervisor generation of the Virtual Machine. Applicable to OS disks only.
+	HyperVGeneration *HyperVGeneration `json:"hyperVGeneration,omitempty"`
+
+	// Describes the gallery image definition purchase plan. This is used by marketplace images.
+	PurchasePlan *ImagePurchasePlan `json:"purchasePlan,omitempty"`
+
+	// The properties describe the recommended machine configuration for this Image Definition. These properties are updatable.
+	Recommended *RecommendedMachineConfiguration `json:"recommended,omitempty"`
+}
+
+// MarshalJSON implements the json.Marshaller interface for type CommunityGalleryImageProperties.
+func (c CommunityGalleryImageProperties) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]interface{})
+	populate(objectMap, "disallowed", c.Disallowed)
+	populate(objectMap, "endOfLifeDate", (*timeRFC3339)(c.EndOfLifeDate))
+	populate(objectMap, "features", c.Features)
+	populate(objectMap, "hyperVGeneration", c.HyperVGeneration)
+	populate(objectMap, "identifier", c.Identifier)
+	populate(objectMap, "osState", c.OSState)
+	populate(objectMap, "osType", c.OSType)
+	populate(objectMap, "purchasePlan", c.PurchasePlan)
+	populate(objectMap, "recommended", c.Recommended)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type CommunityGalleryImageProperties.
+func (c *CommunityGalleryImageProperties) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return err
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "disallowed":
+			err = unpopulate(val, &c.Disallowed)
+			delete(rawMsg, key)
+		case "endOfLifeDate":
+			var aux timeRFC3339
+			err = unpopulate(val, &aux)
+			c.EndOfLifeDate = (*time.Time)(&aux)
+			delete(rawMsg, key)
+		case "features":
+			err = unpopulate(val, &c.Features)
+			delete(rawMsg, key)
+		case "hyperVGeneration":
+			err = unpopulate(val, &c.HyperVGeneration)
+			delete(rawMsg, key)
+		case "identifier":
+			err = unpopulate(val, &c.Identifier)
+			delete(rawMsg, key)
+		case "osState":
+			err = unpopulate(val, &c.OSState)
+			delete(rawMsg, key)
+		case "osType":
+			err = unpopulate(val, &c.OSType)
+			delete(rawMsg, key)
+		case "purchasePlan":
+			err = unpopulate(val, &c.PurchasePlan)
+			delete(rawMsg, key)
+		case "recommended":
+			err = unpopulate(val, &c.Recommended)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+// CommunityGalleryImageVersion - Specifies information about the gallery image version that you want to create or update.
+type CommunityGalleryImageVersion struct {
+	PirCommunityGalleryResource
+	// Describes the properties of a gallery image version.
+	Properties *CommunityGalleryImageVersionProperties `json:"properties,omitempty"`
+}
+
+// CommunityGalleryImageVersionProperties - Describes the properties of a gallery image version.
+type CommunityGalleryImageVersionProperties struct {
+	// The end of life date of the gallery image version Definition. This property can be used for decommissioning purposes. This property is updatable.
+	EndOfLifeDate *time.Time `json:"endOfLifeDate,omitempty"`
+
+	// The published date of the gallery image version Definition. This property can be used for decommissioning purposes. This property is updatable.
+	PublishedDate *time.Time `json:"publishedDate,omitempty"`
+}
+
+// MarshalJSON implements the json.Marshaller interface for type CommunityGalleryImageVersionProperties.
+func (c CommunityGalleryImageVersionProperties) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]interface{})
+	populate(objectMap, "endOfLifeDate", (*timeRFC3339)(c.EndOfLifeDate))
+	populate(objectMap, "publishedDate", (*timeRFC3339)(c.PublishedDate))
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type CommunityGalleryImageVersionProperties.
+func (c *CommunityGalleryImageVersionProperties) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return err
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "endOfLifeDate":
+			var aux timeRFC3339
+			err = unpopulate(val, &aux)
+			c.EndOfLifeDate = (*time.Time)(&aux)
+			delete(rawMsg, key)
+		case "publishedDate":
+			var aux timeRFC3339
+			err = unpopulate(val, &aux)
+			c.PublishedDate = (*time.Time)(&aux)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+// CommunityGalleryImageVersionsGetOptions contains the optional parameters for the CommunityGalleryImageVersions.Get method.
+type CommunityGalleryImageVersionsGetOptions struct {
+	// placeholder for future optional parameters
+}
+
+// CommunityGalleryImagesGetOptions contains the optional parameters for the CommunityGalleryImages.Get method.
+type CommunityGalleryImagesGetOptions struct {
+	// placeholder for future optional parameters
+}
+
 // ComputeOperationListResult - The List Compute Operation operation response.
 type ComputeOperationListResult struct {
 	// READ-ONLY; The list of compute operations
@@ -1800,6 +1972,7 @@ type DedicatedHostsListByHostGroupOptions struct {
 // Minimum api-version: 2015-06-15.
 type DiagnosticsProfile struct {
 	// Boot Diagnostics is a debugging feature which allows you to view Console Output and Screenshot to diagnose VM status.
+	// NOTE: If storageUri is being specified then ensure that the storage account is in the same region and subscription as the VM.
 	// You can easily view the output of your console log.
 	// Azure also enables you to see a screenshot of the VM from the hypervisor.
 	BootDiagnostics *BootDiagnostics `json:"bootDiagnostics,omitempty"`
@@ -5045,6 +5218,21 @@ type PatchSettings struct {
 	PatchMode *WindowsVMGuestPatchMode `json:"patchMode,omitempty"`
 }
 
+// PirCommunityGalleryResource - Base information about the community gallery resource in pir.
+type PirCommunityGalleryResource struct {
+	// The identifier information of community gallery.
+	Identifier *CommunityGalleryIdentifier `json:"identifier,omitempty"`
+
+	// READ-ONLY; Resource location
+	Location *string `json:"location,omitempty" azure:"ro"`
+
+	// READ-ONLY; Resource name
+	Name *string `json:"name,omitempty" azure:"ro"`
+
+	// READ-ONLY; Resource type
+	Type *string `json:"type,omitempty" azure:"ro"`
+}
+
 // PirResource - The Resource model definition.
 type PirResource struct {
 	// READ-ONLY; Resource location
@@ -5889,11 +6077,11 @@ type RestorePointProperties struct {
 	// List of disk resource ids that the customer wishes to exclude from the restore point. If no disks are specified, all disks will be included.
 	ExcludeDisks []*APIEntityReference `json:"excludeDisks,omitempty"`
 
+	// Gets the creation time of the restore point.
+	TimeCreated *time.Time `json:"timeCreated,omitempty"`
+
 	// READ-ONLY; Gets the consistency mode for the restore point. Please refer to https://aka.ms/RestorePoints for more details.
 	ConsistencyMode *ConsistencyModeTypes `json:"consistencyMode,omitempty" azure:"ro"`
-
-	// READ-ONLY; Gets the provisioning details set by the server during Create restore point operation.
-	ProvisioningDetails *RestorePointProvisioningDetails `json:"provisioningDetails,omitempty" azure:"ro"`
 
 	// READ-ONLY; Gets the provisioning state of the restore point.
 	ProvisioningState *string `json:"provisioningState,omitempty" azure:"ro"`
@@ -5907,39 +6095,14 @@ func (r RestorePointProperties) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]interface{})
 	populate(objectMap, "consistencyMode", r.ConsistencyMode)
 	populate(objectMap, "excludeDisks", r.ExcludeDisks)
-	populate(objectMap, "provisioningDetails", r.ProvisioningDetails)
 	populate(objectMap, "provisioningState", r.ProvisioningState)
 	populate(objectMap, "sourceMetadata", r.SourceMetadata)
+	populate(objectMap, "timeCreated", (*timeRFC3339)(r.TimeCreated))
 	return json.Marshal(objectMap)
 }
 
-// RestorePointProvisioningDetails - Restore Point Provisioning details.
-type RestorePointProvisioningDetails struct {
-	// Gets the creation time of the restore point.
-	CreationTime *time.Time `json:"creationTime,omitempty"`
-
-	// Gets the status of the Create restore point operation.
-	StatusCode *int32 `json:"statusCode,omitempty"`
-
-	// Gets the status message of the Create restore point operation.
-	StatusMessage *string `json:"statusMessage,omitempty"`
-
-	// Gets the total size of the data in all the disks which are part of the restore point.
-	TotalUsedSizeInBytes *int64 `json:"totalUsedSizeInBytes,omitempty"`
-}
-
-// MarshalJSON implements the json.Marshaller interface for type RestorePointProvisioningDetails.
-func (r RestorePointProvisioningDetails) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
-	populate(objectMap, "creationTime", (*timeRFC3339)(r.CreationTime))
-	populate(objectMap, "statusCode", r.StatusCode)
-	populate(objectMap, "statusMessage", r.StatusMessage)
-	populate(objectMap, "totalUsedSizeInBytes", r.TotalUsedSizeInBytes)
-	return json.Marshal(objectMap)
-}
-
-// UnmarshalJSON implements the json.Unmarshaller interface for type RestorePointProvisioningDetails.
-func (r *RestorePointProvisioningDetails) UnmarshalJSON(data []byte) error {
+// UnmarshalJSON implements the json.Unmarshaller interface for type RestorePointProperties.
+func (r *RestorePointProperties) UnmarshalJSON(data []byte) error {
 	var rawMsg map[string]json.RawMessage
 	if err := json.Unmarshal(data, &rawMsg); err != nil {
 		return err
@@ -5947,19 +6110,22 @@ func (r *RestorePointProvisioningDetails) UnmarshalJSON(data []byte) error {
 	for key, val := range rawMsg {
 		var err error
 		switch key {
-		case "creationTime":
+		case "consistencyMode":
+			err = unpopulate(val, &r.ConsistencyMode)
+			delete(rawMsg, key)
+		case "excludeDisks":
+			err = unpopulate(val, &r.ExcludeDisks)
+			delete(rawMsg, key)
+		case "provisioningState":
+			err = unpopulate(val, &r.ProvisioningState)
+			delete(rawMsg, key)
+		case "sourceMetadata":
+			err = unpopulate(val, &r.SourceMetadata)
+			delete(rawMsg, key)
+		case "timeCreated":
 			var aux timeRFC3339
 			err = unpopulate(val, &aux)
-			r.CreationTime = (*time.Time)(&aux)
-			delete(rawMsg, key)
-		case "statusCode":
-			err = unpopulate(val, &r.StatusCode)
-			delete(rawMsg, key)
-		case "statusMessage":
-			err = unpopulate(val, &r.StatusMessage)
-			delete(rawMsg, key)
-		case "totalUsedSizeInBytes":
-			err = unpopulate(val, &r.TotalUsedSizeInBytes)
+			r.TimeCreated = (*time.Time)(&aux)
 			delete(rawMsg, key)
 		}
 		if err != nil {
@@ -11278,7 +11444,7 @@ type VirtualMachinesBeginDeallocateOptions struct {
 
 // VirtualMachinesBeginDeleteOptions contains the optional parameters for the VirtualMachines.BeginDelete method.
 type VirtualMachinesBeginDeleteOptions struct {
-	// Optional parameter to force delete virtual machines.(Feature in Preview)
+	// Optional parameter to force delete virtual machines.
 	ForceDeletion *bool
 }