+++ sdk/resourcemanager/compute/armcompute/zz_generated_response_types.go
@@ -10,23 +10,10 @@ package armcompute
 
 import (
 	"context"
+	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
 	"net/http"
 	"time"
-
-	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
 )
-
-// AvailabilitySetsCreateOrUpdateResponse contains the response from method AvailabilitySets.CreateOrUpdate.
-type AvailabilitySetsCreateOrUpdateResponse struct {
-	AvailabilitySetsCreateOrUpdateResult
-	// RawResponse contains the underlying HTTP response.
-	RawResponse *http.Response
-}
-
-// AvailabilitySetsCreateOrUpdateResult contains the result from method AvailabilitySets.CreateOrUpdate.
-type AvailabilitySetsCreateOrUpdateResult struct {
-	AvailabilitySet
-}
 
 // AvailabilitySetsDeleteResponse contains the response from method AvailabilitySets.Delete.
 type AvailabilitySetsDeleteResponse struct {
@@ -1140,6 +1127,42 @@ type CloudServicesUpdateResponse struct {
 // CloudServicesUpdateResult contains the result from method CloudServices.Update.
 type CloudServicesUpdateResult struct {
 	CloudService
+}
+
+// CommunityGalleriesGetResponse contains the response from method CommunityGalleries.Get.
+type CommunityGalleriesGetResponse struct {
+	CommunityGalleriesGetResult
+	// RawResponse contains the underlying HTTP response.
+	RawResponse *http.Response
+}
+
+// CommunityGalleriesGetResult contains the result from method CommunityGalleries.Get.
+type CommunityGalleriesGetResult struct {
+	CommunityGallery
+}
+
+// CommunityGalleryImageVersionsGetResponse contains the response from method CommunityGalleryImageVersions.Get.
+type CommunityGalleryImageVersionsGetResponse struct {
+	CommunityGalleryImageVersionsGetResult
+	// RawResponse contains the underlying HTTP response.
+	RawResponse *http.Response
+}
+
+// CommunityGalleryImageVersionsGetResult contains the result from method CommunityGalleryImageVersions.Get.
+type CommunityGalleryImageVersionsGetResult struct {
+	CommunityGalleryImageVersion
+}
+
+// CommunityGalleryImagesGetResponse contains the response from method CommunityGalleryImages.Get.
+type CommunityGalleryImagesGetResponse struct {
+	CommunityGalleryImagesGetResult
+	// RawResponse contains the underlying HTTP response.
+	RawResponse *http.Response
+}
+
+// CommunityGalleryImagesGetResult contains the result from method CommunityGalleryImages.Get.
+type CommunityGalleryImagesGetResult struct {
+	CommunityGalleryImage
 }
 
 // DedicatedHostGroupsCreateOrUpdateResponse contains the response from method DedicatedHostGroups.CreateOrUpdate.